Added compatibility mode to `mkrlen`. Fixed null terminator not being written by `mkrlen`.
The compatibility mode is necessary because many rlens are formatted weirdly, and I couldn't find any algorithm to follow that would result in their format.
Also only figured out it has a null terminator because without it there's a map part that's a multiple of `0x800` and it was missing a sector.