Started work on remake.
This includes support for building a separate iso with a new executable built exclusively from a rust.
The remake will not be binary perfect, and instead will be focused on replicating the base game in a high-level language with the intent of being easily moddable.
The current iso will not be removed. Instead it will live alongside the remake.