Merged several rust tools into a single package with multiple binaries.
This helps with sharing the configuration files between the `mk*` and `un*` tools.
The naming of things in `{mk,un}lang_file` is a bit weird because we don't really emit/consume a config, per se. It's just the lang file serialized to toml.
This should be changed in the future to contain a config with a reference to the lang toml file.